requirements.txt
distlib==0.3.4
filelock==3.6.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
tabulate==0.8.9
virtualenv==20.14.1
xmltodict==0.12.0